Generate placeholders in a background task
Insteading of generating the placeholders in the same thread, pass them to a background task so they can be processed async. This is better for the performance when uploading images. I also moved to the `updated` rather than `generate` hook, as really I think symnatically it makes a lot more sense.

I didn't go this out of the blue, as it happens, generating placeholders on the update/generate step can be problematic. This is because the attachment metadata is not actually updated in the DB yet, so things that hook into `get_attachment_image_src` won't have the attachment sizes properly available. This was an issue when integrating Tachyon with Gaussholder.