Support external urls
Because we use `wp_get_attachment_image_src` to get the image path (by doing a str_replace), external urls can also be returned. This happens in the case of using say, Tachyon. That's not actually a bad thing though, in these cases, we don't want to make a gaussholder placeholder of the image on disk, we want to make one of the external image. This adds support for such scenarios.