Don't no-op when there's no srcset
For some reason, we currently bail if there's no srcset set on an image, but that is a quite legitimate thing to do. For example, thumbnail size images that are added to posts. These are unlikely to have a srcset, as there are no other default image sizes that are cropped.