Use MutationObserver to update image dimensions when it renders
The dimension calculations were calculating image dimensions before the
image is rendered on the page, which gives incorrect element dimensions,
leading to a misproportioned image.

This PR attempts to fix that issue by recalculating dimensions on a
Mutation event, when the src is first replaced, rather than immediately
on setup, to better calculate the space the <img> element takes up in
the document. (I'm torn between this approach and just doing it in a
timeout... I think the timeout would be better performance, but at the
risk of not being as accurate.)